feat: add responsive drawer and navigation to dashboard page with new icons
